Added 2 random targets to show how hovering works
